% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rejection_Samplr.R
\name{rejection_samplr}
\alias{rejection_samplr}
\title{Add together two numbers}
\usage{
rejection_samplr(n, pdf, a, b, C)
}
\arguments{
\item{n}{the number of samples}

\item{pdf}{a function that is the pdf of the random variable that you wish to sample from}

\item{a}{a numeric that is the lower bound of the random variable you wish to sample from}

\item{b}{a numeric that is the upper bound of the random variable you wish to sample from, so P(a ≤ X ≤ b) = 1}

\item{C}{a numeric that is such that f(x) ≤ C for all values of x}
}
\value{
a random sample of size n from the rv with pdf provided
}
\description{
This is code that performs rejection sampling on a continuous random variable
}
\examples{
n/a
}
