% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2DRejectionSampling.R
\name{samplr2D}
\alias{samplr2D}
\title{2D Rejection Sampler on Square}
\usage{
samplr2D(n, jpdf, a, b, C)
}
\arguments{
\item{n}{the number of samples}

\item{jpdf}{a function that is the joint pdf of the random variable from which you want to sample. Should use x and y.}

\item{a}{a numeric that is the lower bound of the random variable you wish to sample from}

\item{b}{a numeric that is the upper bound of the random variable you wish to sample from}

\item{C}{a numeric that is such that f(x) ≤ C for all values of x and y}
}
\value{
a data frame with x and y as the column names and the rows filled with the accepted values
}
\description{
This is code that performs rejection sampling on a continuous random variable
}
\examples{
## sampling from two unif(0,1) rv's
 unif_jpdf <- function(x,y, min = 0, max = 1){
 x <- dunif(x, min = min, max = max)
 y <- dunif(y, min = min, max = max)
 x*y
 }
 test <- samplr2D(n = 2000, jpdf = unif_jpdf, a = 0, b = 1, C = 1)
}
